Extract productSlugs logic to 'useProductSlugs' custom hook.
