+++ client/my-sites/plans/jetpack-plans/product-store/index.tsx
@@ -1,11 +1,8 @@
-import { useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import IntroPricingBanner from 'calypso/components/jetpack/intro-pricing-banner';
 import StoreFooter from 'calypso/jetpack-connect/store-footer';
-import { getSitePlan } from 'calypso/state/sites/selectors';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
-import { getPlansToDisplay, getProductsToDisplay } from '../product-grid/utils';
-import useGetPlansGridProducts from '../use-get-plans-grid-products';
+import useProductSlugs from './hooks/use-product-slugs';
 import { JetpackFree } from './jetpack-free';
 import { Recommendations } from './recommendations';
 import { UserLicensesDialog } from './user-licenses-dialog';
@@ -19,24 +16,7 @@ const ProductStore: React.FC< ProductStoreProps > = ( {
 	duration,
 } ) => {
 	const siteId = useSelector( getSelectedSiteId );
-	const currentPlan = useSelector( ( state ) => getSitePlan( state, siteId ) );
-	const currentPlanSlug = currentPlan?.product_slug || null;
-	const { availableProducts, purchasedProducts, includedInPlanProducts } =
-		useGetPlansGridProducts( siteId );
-
-	const productSlugs = useMemo(
-		() =>
-			[
-				...getProductsToDisplay( {
-					duration,
-					availableProducts,
-					purchasedProducts,
-					includedInPlanProducts,
-				} ),
-				...getPlansToDisplay( { duration, currentPlanSlug } ),
-			].map( ( { productSlug } ) => productSlug ),
-		[ duration, availableProducts, purchasedProducts, includedInPlanProducts, currentPlanSlug ]
-	);
+	const productSlugs = useProductSlugs( siteId, duration );
 
 	return (
 		<div className="jetpack-product-store">
